iMove access determination tests to a separate test class
